pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-ai-app"
version = "0.1.0"
description = "企业级生成式AI应用Web服务框架"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AI Framework Team" }]
keywords = ["fastapi", "crewai", "ai", "enterprise"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "prometheus-fastapi-instrumentator>=7.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
    "aiomysql>=0.2.0",
    "redis>=5.0.0",
    "alembic>=1.13",
    "aiofiles>=24.0",
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[project.urls]
Documentation = "https://github.com/kid0317/fastapi_base"
Repository = "https://github.com/kid0317/fastapi_base"
Homepage = "https://github.com/kid0317/fastapi_base"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
# 集成测试时应用日志（含 trace）需在控制台显示，默认不捕获 stdout
addopts = "-s"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
